requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==4.1.1
djangorestframework==3.13.1
docopt==0.6.2
git-ignore==1.0.2
gitignore==0.0.8
idna==3.3
nr.gitignore==1.0.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
